9501 - adding more coverage and splitting up the large test file into two smaller modules
